front/src/app/features/auth/components/login/login.component.ts
import { Component } from '@angular/core';
import { FormBuilder, Validators } from '@angular/forms';
import { Router } from '@angular/router';
import { SessionInformation } from 'src/app/interfaces/sessionInformation.interface';
import { SessionService } from 'src/app/services/session.service';
import { LoginRequest } from '../../interfaces/loginRequest.interface';
import { AuthService } from '../../services/auth.service';

@Component({
  selector: 'app-login',
  templateUrl: './login.component.html',
  styleUrls: ['./login.component.scss'],
})
export class LoginComponent {
  public hide = true;
  public onError = false;

  public form = this.fb.group({
    email: ['', [Validators.required, Validators.email]],
    password: ['', [Validators.required, Validators.min(3)]],
  });

  constructor(
    private authService: AuthService,
    private fb: FormBuilder,
    private router: Router,
    private sessionService: SessionService
  ) {}

  public submit(): void {
    const loginRequest = this.form.value as LoginRequest;

    console.log('Submitting login request:', loginRequest);

    this.authService.login(loginRequest).subscribe({
      next: (response: SessionInformation) => {
        console.log('Login successful. Session Information:', response);
        this.sessionService.logIn(response);
        this.router.navigate(['/sessions']);
      },
      error: (error) => {
        console.error('Error during login:', error);
        this.onError = true;
      },
    });
  }
}
